Fixed bug where you could still do actions when logged out

+++ server/controllers/comment.js
@@ -5,6 +5,8 @@ const { body } = require("express-validator");
 const router = require("express").Router();
 
 router.post("/", body("body").trim(), async (req, res) => {
+  if (req.user === undefined) return res.json(false);
+
   const newComment = await Comment.create({
     author: req.user._id,
     body: req.body.body,
@@ -14,14 +16,16 @@ router.post("/", body("body").trim(), async (req, res) => {
   post.comments.unshift(newComment._id);
   await post.save();
 
-  res.end();
+  res.json(true);
 });
 
 router.delete("/:commentId", async (req, res) => {
+  if (req.user === undefined) return res.json(false);
+
   const commentId = req.params.commentId;
   await Comment.findByIdAndDelete(commentId).exec();
 
-  res.end();
+  res.json(true);
 });
 
 module.exports = router;
